updoots to all the scripts

+++ platform/debian/build_kanidm.sh
@@ -4,16 +4,20 @@
 
 set -e
 
-if [ -z $1 ]; then
+if [ -z "$1" ]; then
     PACKAGE="kanidm"
 else
     PACKAGE="$1"
 fi
 
+if [ ! -d "./platform/debian/${PACKAGE}" ]; then
+    echo "Can't find packaging files for ${PACKAGE}"
+    exit 1
+fi
 
 echo "Building ${PACKAGE}"
 
-
+# gotta do the sudo thing in github actions, or... in general
 if [ "$(whoami)" != "root" ]; then
     SUDO="sudo "
 else
@@ -22,22 +26,23 @@ fi
 if [ -n "${GITHUB_WORKSPACE}" ]; then
     SOURCE_DIR="${GITHUB_WORKSPACE}"
 else
-    SOURCE_DIR="$HOME/kanidm"
+    SOURCE_DIR="${HOME}/kanidm"
 fi
 BUILD_DIR="$HOME/build"
-PACKAGE_DIR="${BUILD_DIR}/kanidm"
-BINARY_DIR="${PACKAGE_DIR}/usr/local/bin"
 
 
 if [ -z "${SKIP_DEPS}" ]; then
-
-    ${SUDO}./platform/debian/install-deps.sh
+    "${SUDO}./platform/debian/install_deps.sh"
+else
+    echo "SKIP_DEPS configured, skipping install of rust and packages"
 fi
 
+#shellcheck disable=SC1091
+source "$HOME/.cargo/env"
 
 # if we can't find cargo then need to update the path
 if [ "$(which cargo | wc -l)" -eq 0 ]; then
-    if [  -z "$(echo $PATH | grep -o '.cargo/bin')" ]; then
+    if echo "$PATH" | grep -q '.cargo/bin'; then
         echo "Updating path to include local cargo dir"
         export PATH="$HOME/.cargo/bin:$PATH"
     fi
@@ -53,34 +58,19 @@ else
     GIT_HEAD="$(git rev-parse HEAD)"
 fi
 
+# we only want the short commit
 GIT_COMMIT="${GIT_HEAD:0:7}"
-# DATESTR="$(date +%Y%m%d%H%M)"
+DATESTR="$(date +%Y%m%d%H%M)"
 
-# PACKAGE_VERSION="${KANIDM_VERSION}-${DATESTR}-${GIT_COMMIT}"
-PACKAGE_VERSION="${KANIDM_VERSION}-${GIT_COMMIT}"
+PACKAGE_VERSION="${KANIDM_VERSION}-${DATESTR}${GIT_COMMIT}"
 echo "Package Version: ${PACKAGE_VERSION}"
 
-# echo "Building kanidm"
-# make local/kanidm
-
 echo "Updating package dir"
-rm -rf "${BUILD_DIR}/*"
-
-# where the binary will go
-# mkdir -p "${BINARY_DIR}"
-# just debian things
-# mkdir -p "${PACKAGE_DIR}/DEBIAN"
-# cp platform/debian/kanidm/* "${PACKAGE_DIR}/DEBIAN"
+rm -rf "${BUILD_DIR:?}/*"
 
 echo "Copying source files to ${BUILD_DIR}"
 rsync -a \
-    --exclude .git \
     --exclude target \
-    --exclude kanidm_book \
-    --exclude artwork \
-    --exclude pykanidm \
-    --exclude designs \
-    --exclude project_docs \
     "${SOURCE_DIR}" \
     "${BUILD_DIR}/"
 
@@ -120,4 +110,8 @@ debian/rules build
 echo "Packaging ${PACKAGE}"
 fakeroot debian/rules binary
 
+echo "Moving debs to target/"
 find ../ -maxdepth 1 -name '*.deb' -exec mv "{}" "${SOURCE_DIR}/target/" \;
+
+echo "Done, packages:"
+find "${SOURCE_DIR}/target/" -maxdepth 1 -name '*.deb'
